feat: 결함 bbox 풀스택 float(서브픽셀) 통일
엣지: YOLO xywh 정수화 제거, BoundingBox/DefectPayload float, ROI·서브픽셀 경로 정수 슬라이스만 사용.

Spring: DefectDetail·DTO Double + 검증. .NET: DefectRequest/DefectRecord double.

프론트: 검출 좌표 toFixed(4) 표시. sender 5xx 응답 본문 로그, Gemini 일일 쿼터 시 재시도 생략.

+++ backend-dotnet/src/AiFactory.Api/Contracts/InspectionIngestRequest.cs
@@ -69,14 +69,14 @@ public sealed class DefectRequest
     public double Confidence { get; set; }
 
     [JsonPropertyName("bboxX")]
-    public int BboxX { get; set; }
+    public double BboxX { get; set; }
 
     [JsonPropertyName("bboxY")]
-    public int BboxY { get; set; }
+    public double BboxY { get; set; }
 
     [JsonPropertyName("bboxWidth")]
-    public int BboxWidth { get; set; }
+    public double BboxWidth { get; set; }
 
     [JsonPropertyName("bboxHeight")]
-    public int BboxHeight { get; set; }
+    public double BboxHeight { get; set; }
 }
